Prototype Code for Player and Food
Dynamic Camera introduced to follow player and look a bit in front of where the player is looking. Player movement for keyboard added, and flocking code for the food has been devised within this version. The enemy(red cube) is just currently random movements, but the food has been programmed to track it in a basic manner.